+++ src/commonMain/kotlin/experimental/AnyInt.kt
@@ -5,7 +5,8 @@
 
 package kotools.types.experimental
 
-import kotools.types.ExperimentalSinceKotoolsTypes
+import kotools.types.internal.ExperimentalSince
+import kotools.types.internal.KotoolsTypesVersion
 import kotools.types.number.AnyInt
 import kotlin.jvm.JvmSynthetic
 
@@ -23,7 +24,7 @@ import kotlin.jvm.JvmSynthetic
  * Please note that this function is currently not available for Java users.
  */
 @ExperimentalKotoolsTypesApi
-@ExperimentalSinceKotoolsTypes("4.3.2")
+@ExperimentalSince(KotoolsTypesVersion.V4_3_2)
 @JvmSynthetic
 public operator fun AnyInt.unaryMinus(): AnyInt {
     val value: Int = -toInt()